requirements.txt
boto3==1.34.82
cryptography==42.0.5
python-dotenv==1.0.1
python_bcrypt==0.3.2
Requests==2.32.3
